+++ src/GRA.Data/Model/Site.cs
@@ -40,7 +40,6 @@ namespace GRA.Data.Model
 
         public string ExternalEventListUrl { get; set; }
         public string GoogleAnalyticsTrackingId { get; set; }
-        public bool CollectPreregistrationEmails { get; set; }
         [MaxLength(150)]
         public string MetaDescription { get; set; }